commit the Cargo.lock
The official guideline for Rust project was originally not to commit the
Cargo.lock for libraries. However, this has been recently revised to
ensure all developers in a project work in a consistent and reproducible
environment.

Furthermore, `svdtools` is a CLI, for which committing the Cargo.lock
was also encouraged. It will also help packaging `svdtools` via Nix
using naersk, which bases on the Cargo.lock to ensure the
reproducibility of the build.